requirements.txt
pyudev
pynacl
pycryptodome
argon2-cffi
fusepy
customtkinter
